requirements.txt
pydantic==1.6.1
flashtext==2.7
pandas==1.0.3
pymongo==3.10.1
fastapi==0.62.0
protobuf==3.14.0
PyYAML==5.3.1
rasa==2.1.3
uvicorn
google-auth
google-cloud
google-cloud-pubsub
